+++ app/scripts/controllers/gov-request-category-list.js
@@ -5,10 +5,13 @@ function($scope, $location, gov_request_categories, urls, dataProviderService, $
 	$scope.gov_request_categories = gov_request_categories;
 
 	$scope.deleteCategory = function(category){
-		dataProviderService.deleteItem(urls.apiURL(), "/gov-request-categories/" + category.category_id)
- 		.then(function(gov_request_categories){
- 			$route.reload()
- 		})
+		var areYouSure = window.confirm("Are you sure you want to delete this category?");
+		if(areYouSure){
+			dataProviderService.deleteItem(urls.apiURL(), "/gov-request-categories/" + category.category_id)
+	 		.then(function(gov_request_categories){
+	 			$route.reload()
+	 		});
+ 		}
 	}
 }
 ]);
