+++ app/scripts/controllers/report-list.js
@@ -5,10 +5,13 @@ function($scope, $location, reports, urls, dataProviderService, $route){
 	$scope.reports = reports;
 
 	$scope.deleteReport = function(report){
-		dataProviderService.deleteItem(urls.apiURL(), "/transparency-reports/" + report.report_id)
- 		.then(function(reports){
- 			$route.reload()
- 		})
+		var areYouSure = window.confirm("Are you sure you want to delete this report?");
+		if(areYouSure){
+			dataProviderService.deleteItem(urls.apiURL(), "/transparency-reports/" + report.report_id)
+	 		.then(function(reports){
+	 			$route.reload()
+	 		});
+ 		}
 	}
 }
 ]);
